refactor check commands into sep. files
